[Feature] Introduced session id
I added session id which is updated when a destructive change occurs
in the log.
(which happens quite often by running the same program with different
parameter set while keeping chainerboard open)
For this I added detection of destructive changes in log file,
passing around of session id between the host and the front and
thread lock mechanism which provents destructive change to break
on-going communication. (updating is locked until a communication
ends and vice versa).

Resolves: #6